Migrate Java 8 tests out of super source
