+++ user/test/com/google/gwt/dev/jjs/test/defaultmethods/ImplementsWithDefaultMethodAndStaticInitializer.java
