+++ user/test/com/google/gwt/dev/jjs/test/defaultmethods/WithDefaultMethodAndStaticInitializer.java
