+++ user/test/com/google/gwt/dev/jjs/test/package1/SimpleA.java
