+++ user/test/com/google/gwt/dev/jjs/test/package2/SimpleB.java
