+++ user/test/com/google/gwt/dev/jjs/test/package3/SimpleC.java
